added function of showing the current stars user rating that firebase has

+++ app/src/main/java/com/example/what2watch_svitlik_odunsi_f23/ui/quizresults/RecyclerResultsCard.kt
@@ -7,5 +7,5 @@ data class RecyclerResultsCard(
     val startYear: Long,
     val genre: String,
     val averageRating: String,
-    val userRating: Int
+    val userRating: Long
 )
